@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Security;
 using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
 using GoldenAnvil.Utility;
+using GoldenAnvil.Utility.Logging;
 using GoldenAnvil.Utility.Windows.Async;
 using Microsoft.VisualStudio.Threading;
 using Oraculum.Engine;
@@ -14,259 +16,450 @@ namespace Oraculum.Data
 {
 	public static class DataImportUtility
 	{
-		public static async Task<IReadOnlyList<(TableMetadata Metadata, IReadOnlyList<RowData> Rows)>> ImportTablesAsync(TaskStateController state, string path)
+		public static async Task<IReadOnlyList<(TableMetadataDto Metadata, IReadOnlyList<RowDataDto> Rows)>> ImportTablesAsync(TaskStateController state, string filePath)
 		{
+			await state.ToSyncContext();
+
+			var titlesInUse = AppModel.Instance.Data.GetAllTableTitles();
+
 			await state.ToThreadPool();
 
-			var tables = new List<(TableMetadata Metadata, IReadOnlyList<RowData> Rows)>();
+			using var logScope = Log.TimedInfo($"Importing tables from \"{filePath}\"...");
 
-			var titlesInUse = await AppModel.Instance.Data.GetAllTableTitlesAsync(state.CancellationToken).ConfigureAwait(false);
+			string defaultTitle = CreateBestTitleFromFileName(titlesInUse, Path.GetFileNameWithoutExtension(filePath));
 
-			var defaultTitle = CreateBestTitleFromFileName(titlesInUse, Path.GetFileNameWithoutExtension(path));
+			string[]? lines = null;
+			try
+			{
+				lines = await File.ReadAllLinesAsync(filePath, state.CancellationToken).ConfigureAwait(false);
+			}
+			catch (Exception ex) when (ex is IOException or UnauthorizedAccessException or NotSupportedException or SecurityException or ArgumentException)
+			{
+				Log.Info($"Error reading file: {ex}");
+				return new List<(TableMetadataDto Metadata, IReadOnlyList<RowDataDto> Rows)>();
+			}
 
-			var lines = await File.ReadAllLinesAsync(path, state.CancellationToken).ConfigureAwait(false);
+			try
+			{
+				return ParseLines(lines, defaultTitle);
+			}
+			catch (FormatException ex)
+			{
+				Log.Info($"Error parsing file: {ex}");
+				return new List<(TableMetadataDto Metadata, IReadOnlyList<RowDataDto> Rows)>();
+			}
+		}
 
-			List<(int? Number1, int? Number2, string Output, Guid? Next)>? currentRowInfos = null;
-			string currentTitle = defaultTitle;
+		private static IReadOnlyList<(TableMetadataDto Metadata, IReadOnlyList<RowDataDto> Rows)> ParseLines(IReadOnlyList<string> lines, string defaultTitle)
+		{
+			var tables = new List<(TableMetadataDto Metadata, IReadOnlyList<RowDataDto> Rows)>();
+
+			Guid? currentId = null;
+			var currentTitle = defaultTitle;
 			string? currentSource = null;
 			string? currentAuthor = null;
-			Guid? currentId = null;
+			var currentVersion = 1;
+			var currentCreated = DateOnly.FromDateTime(DateTime.Now);
+			var currentModified = currentCreated;
+			string? currentDescription = null;
+			RandomPlan? currentRandomPlan = null;
+			RandomSourceBase? currentRandomSource = null;
 			IReadOnlyList<string> currentGroups = Array.Empty<string>();
-			IReadOnlyList<RandomSourceData> currentRandomPlan = Array.Empty<RandomSourceData>();
+			List<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)>? currentRowInfos = null;
 
-			var isAllWhitespaceRegex = new Regex(@"^\s*$");
 			var parseState = ImportParseState.ReadingMetadata;
-			foreach (var line in lines)
+			foreach (var line in lines.Select(x => x.Trim()))
 			{
-				if (line.Length == 0 || isAllWhitespaceRegex.IsMatch(line))
+				if (line.Length == 0)
 					continue;
 
 				if (line.StartsWith("# "))
 				{
 					if (parseState == ImportParseState.ReadingRows)
 					{
-						CreateTableMetadata(currentId, currentTitle, currentAuthor, currentSource, currentRandomPlan, currentGroups, currentRowInfos!);
-						currentGroups = Array.Empty<string>();
-						currentRandomPlan = Array.Empty<RandomSourceData>();
+						tables.Add(CreateTable(currentId, currentTitle, currentSource, currentAuthor, currentVersion, currentCreated, currentModified, currentDescription, currentRandomPlan, currentGroups, currentRowInfos));
+						currentId = null;
 						currentSource = null;
 						currentAuthor = null;
-						currentId = null;
+						currentVersion = 1;
+						currentCreated = DateOnly.FromDateTime(DateTime.Now);
+						currentModified = currentCreated;
+						currentDescription = null;
+						currentRandomPlan = null;
+						currentRandomSource = null;
+						currentGroups = Array.Empty<string>();
 						currentRowInfos = null;
 						parseState = ImportParseState.ReadingMetadata;
 					}
-					currentTitle = line[2..].Trim();
+					currentTitle = line[2..];
 				}
 				else
 				{
 					if (parseState == ImportParseState.ReadingMetadata)
 					{
-						if (line.StartsWith("Groups:", StringComparison.OrdinalIgnoreCase))
+						if (line.StartsWith("ID:", StringComparison.OrdinalIgnoreCase))
 						{
-							currentGroups = line[7..]
-								.Split(" > ", StringSplitOptions.TrimEntries | StringSplitOptions.RemoveEmptyEntries)
-								.ToList();
+							currentId = Guid.Parse(line[3..]);
 						}
 						else if (line.StartsWith("Source:", StringComparison.OrdinalIgnoreCase))
 						{
-							currentSource = line[7..].Trim();
+							currentSource = line[7..];
 						}
 						else if (line.StartsWith("Author:", StringComparison.OrdinalIgnoreCase))
 						{
-							currentAuthor = line[7..].Trim();
+							currentAuthor = line[7..];
 						}
-						else if (line.StartsWith("ID:", StringComparison.OrdinalIgnoreCase))
+						else if (line.StartsWith("Version:", StringComparison.OrdinalIgnoreCase))
 						{
-							currentId = Guid.Parse(line[3..].Trim());
+							if (int.TryParse(line[8..], NumberStyles.Integer, CultureInfo.InvariantCulture, out var version))
+								currentVersion = version;
+							else
+								throw new FormatException($"Version must be an integer: {line}");
+						}
+						else if (line.StartsWith("Created:", StringComparison.OrdinalIgnoreCase))
+						{
+							var dateValue = line[8..].Trim();
+							var yearRegex = new Regex(@"^\d{4}$");
+							if (yearRegex.IsMatch(dateValue))
+								currentCreated = new DateOnly(int.Parse(dateValue), 1, 1);
+							else if (DateOnly.TryParse(line[8..], CultureInfo.InvariantCulture, out var created))
+								currentCreated = created;
+							else
+								throw new FormatException($"Created date could not be parsed: {line}");
+						}
+						else if (line.StartsWith("Modified:", StringComparison.OrdinalIgnoreCase))
+						{
+							if (DateOnly.TryParse(line[9..], CultureInfo.InvariantCulture, out var modified))
+								currentModified = modified;
+							else
+								throw new FormatException($"Updated date could not be parsed: {line}");
+						}
+						else if (line.StartsWith("Description:", StringComparison.OrdinalIgnoreCase))
+						{
+							currentDescription = line[12..];
 						}
 						else if (line.StartsWith("Random:", StringComparison.OrdinalIgnoreCase))
 						{
-							currentRandomPlan = ParseRandomPlan(line[7..].Trim());
+							currentRandomPlan = ParseRandomPlan(line[7..]);
+							currentRandomSource = RandomSourceBase.Create(currentRandomPlan);
+						}
+						else if (line.StartsWith("Groups:", StringComparison.OrdinalIgnoreCase))
+						{
+							currentGroups = line[7..]
+								.Split(">", StringSplitOptions.TrimEntries | StringSplitOptions.RemoveEmptyEntries)
+								.ToList();
 						}
 						else
 						{
-							currentRowInfos = new List<(int? Number1, int? Number2, string Output, Guid? Next)>();
+							currentRowInfos = new List<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)>();
 							parseState = ImportParseState.ReadingRows;
 						}
 					}
 
 					if (parseState == ImportParseState.ReadingRows)
-						currentRowInfos!.Add(CreateRowInfo(line));
+						currentRowInfos!.Add(CreateRowInfo(line, currentRandomSource));
 				}
 			}
-			if (currentRowInfos is not null)
-				CreateTableMetadata(currentId, currentTitle, currentAuthor, currentSource, currentRandomPlan, currentGroups, currentRowInfos);
-
-			void CreateTableMetadata(Guid? id, string title, string? author, string? source, IReadOnlyList<RandomSourceData> randomPlan, IReadOnlyList<string> groups, List<(int? Number1, int? Number2, string Output, Guid? Next)> rowInfos)
-			{
-				var rowType = GuessRowType(rowInfos);
-
-				var (randomSource, rows) = rowType switch
-				{
-					RowKind.NoNumbers => CreateNoNumbersRowDatas(rowInfos),
-					RowKind.MixedRanges => CreateMixedRangesRowDatas(rowInfos),
-					RowKind.Weighted => CreateWeightedRowDatas(rowInfos),
-					_ => throw new InvalidOperationException("Unknown row kind"),
-				};
-
-				if (randomPlan.Count == 0)
-					randomPlan = new[] { randomSource };
-				else if (randomPlan[0] != randomSource)
-					throw new FormatException("Calculated random source does not match the expected random plan.");
-
-				var metadata = new TableMetadata
-				{
-					Id = id ?? Guid.NewGuid(),
-					Title = title,
-					Source = source,
-					Author = author,
-					Version = 1,
-					Created = DateTime.Now,
-					Modified = DateTime.Now,
-					RandomPlan = randomPlan,
-					Groups = groups,
-				};
-
-				tables.Add((metadata, rows));
-			}
+			if (currentTitle is not null)
+				tables.Add(CreateTable(currentId, currentTitle, currentSource, currentAuthor, currentVersion, currentCreated, currentModified, currentDescription, currentRandomPlan, currentGroups, currentRowInfos));
 
 			return tables;
 		}
 
+		private static (TableMetadataDto Metadata, IReadOnlyList<RowDataDto> Rows) CreateTable(Guid? id, string title, string? source, string? author, int version, DateOnly created, DateOnly modified, string? description, RandomPlan? randomPlan, IReadOnlyList<string> groups, IReadOnlyList<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)>? rowInfos)
+		{
+			if (rowInfos is null)
+				throw new FormatException($"Table is missing rows: {title}");
+
+			var transformedRows = rowInfos;
+
+			if (randomPlan is null)
+			{
+				randomPlan = GuessRandomPlan(rowInfos);
+				if (rowInfos[0].GuessedConfigs is null)
+				{
+					transformedRows = ApplyRowWeights().AsReadOnlyList();
+
+					IEnumerable<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)> ApplyRowWeights()
+					{
+						var currentWeight = 0;
+						foreach (var rowInfo in rowInfos)
+						{
+							currentWeight += rowInfo.Value1.Values[0];
+							yield return (new DieValue(currentWeight), null, null, rowInfo.Output);
+						}
+					}
+				}
+				else if (rowInfos[0].Value1.Values.Count == 1 && randomPlan.Configurations.Count > 1)
+				{
+					transformedRows = ApplySequenceTransform().AsReadOnlyList();
+
+					IEnumerable<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)> ApplySequenceTransform()
+					{
+						foreach (var rowInfo in rowInfos)
+						{
+							var value1 = new DieValue(DigitCollection.Create(rowInfo.Value1.Values[0])
+								.Select(x => x)
+								.AsReadOnlyList());
+							RandomValueBase? value2 = null;
+							if (rowInfo.Value2 is not null)
+							{
+								value2 = new DieValue(DigitCollection.Create(rowInfo.Value2.Values[0])
+									.Select(x => x)
+									.AsReadOnlyList());
+							}
+							yield return (value1, value2, rowInfo.GuessedConfigs, rowInfo.Output);
+						}
+					}
+				}
+			}
+
+			ValidateRows(transformedRows, randomPlan);
+			var rows = CreateRows(transformedRows);
+
+			var metadata = new TableMetadataDto(
+				tableId: id ?? Guid.NewGuid(),
+				title: title,
+				source: source,
+				author: author,
+				version: 1,
+				created: created,
+				modified: modified,
+				description: description,
+				randomPlan: randomPlan!,
+				groups: groups
+			);
+
+			return (metadata, rows);
+		}
+
+		private static void ValidateRows(IReadOnlyList<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)> rows, RandomPlan randomPlan)
+		{
+			var values = new HashSet<RandomValueBase>();
+			foreach (var row in rows)
+			{
+				if (row.Value2 is null)
+				{
+					if (!values.Add(row.Value1))
+						throw new FormatException($"Duplicate row value: {row.Value1}");
+				}
+				else
+				{
+					foreach (var value in row.Value1.EnumerateTo(row.Value2, randomPlan.Configurations))
+					{
+						if (!values.Add(value))
+							throw new FormatException($"Duplicate row value: {value}");
+					}
+				}
+			}
+		}
+
+		private static RandomPlan GuessRandomPlan(IReadOnlyList<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)> rowInfos)
+		{
+			var valueKind = rowInfos[0].Value1.Kind;
+			if (rowInfos.Any(x => (x.Value1?.Kind ?? RandomValueKind.Die) != valueKind))
+				throw new FormatException("All Row values must be the same kind.");
+
+			RandomSourceKind sourceKind;
+			IReadOnlyList<int>? configs;
+			var configCount = rowInfos[0].GuessedConfigs?.Count;
+			if (configCount is null)
+			{
+				if (rowInfos.Any(x => x.GuessedConfigs is not null))
+					throw new FormatException("All Row values must have the same number of configurations.");
+				sourceKind = RandomSourceKind.DiceSequence;
+				configs = new[] { rowInfos.Sum(x => x.Value1.Values[0]) };
+			}
+			else
+			{
+				if (configCount is not null && rowInfos.Any(x => x.GuessedConfigs?.Count != configCount))
+					throw new FormatException("All Row values must have the same number of configurations.");
+
+				configs = valueKind switch
+				{
+					RandomValueKind.Card => CardUtility.MergeConfigurations(rowInfos.Select(x => x.GuessedConfigs!)),
+					RandomValueKind.Die => DieUtility.MergeConfigurations(rowInfos.Select(x => x.GuessedConfigs!)),
+					_ => throw new NotImplementedException("Unknown random value kind"),
+				};
+				if (configs is null)
+					throw new FormatException("Row configurations are not compatible.");
+
+				sourceKind = valueKind == RandomValueKind.Die ? RandomSourceKind.DiceSequence : RandomSourceKind.CardSequence;
+				var minValue = rowInfos.Min(x => x.Value1.Values[0]);
+				if (sourceKind == RandomSourceKind.DiceSequence && configs.Count == 1 && minValue > 1)
+				{
+					DigitCollection minDigits = minValue;
+					DigitCollection maxDigits = rowInfos.Max(x => (x.Value2 ?? x.Value1).Values[0]);
+
+					if (minDigits.Count == maxDigits.Count)
+					{
+						var hasValidMinValue = minDigits.All(x => x == 1);
+						var hasValidMaxValue = maxDigits.All(x => x == maxDigits[0]);
+						var hasCorrectRowCount = ((int) Math.Pow(maxDigits[0], maxDigits.Count)) == rowInfos.Count;
+						if (hasValidMinValue && hasValidMaxValue && hasCorrectRowCount)
+							return new RandomPlan(sourceKind, maxDigits.ToArray());
+					}
+
+					throw new FormatException("Unable to guess random plan from rows.");
+				}
+			}
+
+			return new RandomPlan(sourceKind, configs);
+		}
+
+		private static IReadOnlyList<RowDataDto> CreateRows(IEnumerable<(RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output)> rowInfos)
+		{
+			return rowInfos
+				.Select(x => new RowDataDto(x.Value1.Values, (x.Value2 ?? x.Value1).Values, x.Output))
+				.AsReadOnlyList();
+		}
+
 		private static string CreateBestTitleFromFileName(HashSet<string> tableTitles, string fileName)
 		{
-			var title = StringUtility.GetWordsFromCamelCase(fileName).Join(" ");
+			string title = StringUtility.GetWordsFromCamelCase(fileName).Join(" ").Replace('_', ' ');
 			return TitleUtility.GetUniqueTitle(title, tableTitles);
 		}
 
-		private static IReadOnlyList<RandomSourceData> ParseRandomPlan(string input)
+		private static RandomPlan ParseRandomPlan(string input)
 		{
-			var parts = input.Split(";", StringSplitOptions.TrimEntries | StringSplitOptions.RemoveEmptyEntries);
-			var regex = new Regex(@"(\d+)d(\d+)");
-			return parts
-				.Select(text =>
+			var kind = RandomSourceKind.DiceSequence;
+
+			var isSum = input.Contains("+");
+			var parts = input.Split(isSum ? "+" : ";", StringSplitOptions.TrimEntries | StringSplitOptions.RemoveEmptyEntries);
+
+			var diceConfigRegex = new Regex(@"(\d*)d(\d+)");
+			var cardConfigRegex = new Regex($"({EnumUtility.Values<CardSourceConfiguration>().Select(x => x.ToString()).Join("|")})");
+			if (parts.All(x => diceConfigRegex.IsMatch(x)))
+				kind = isSum ? RandomSourceKind.DiceSum : RandomSourceKind.DiceSequence;
+			else if (parts.All(x => cardConfigRegex.IsMatch(x)) && !isSum)
+				kind = RandomSourceKind.CardSequence;
+			else
+				throw new FormatException($"Random does not have the expected format: {input}");
+
+			var configurations = parts
+				.SelectMany(part =>
 				{
-					var match = regex.Match(text);
-					if (!match.Success)
-						throw new FormatException("Random plan input does not have the expected format.");
-					var count = int.Parse(match.Groups[1].Value);
-					var sides = int.Parse(match.Groups[2].Value);
-					return new RandomSourceData(RandomSourceKind.Die, Enumerable.Range(1, count).Select(x => sides).AsReadOnlyList());
-				})
-				.AsReadOnlyList();
-		}
-
-		private static (int? Number1, int? Number2, string Output, Guid? next) CreateRowInfo(string line)
-		{
-			var regex = new Regex(@"^(?>\s*(\d+)\s+)?(?>(\d+)\s+)?\.?\s*(.*)$");
-			var match = regex.Match(line);
-			var number1 = match.Groups[1].Success ? StringUtility.TryParse<int>(match.Groups[1].Value) : null;
-			var number2 = match.Groups[2].Success ? StringUtility.TryParse<int>(match.Groups[2].Value) : null;
-			if (number1 == 0)
-				number1 = 100;
-			if (number2 == 0)
-				number2 = 100;
-			var output = match.Groups[3].Value;
-			var nextRegex = new Regex(@"(\s*\[(.*)\])?$");
-			var nextMatch = nextRegex.Match(output);
-			Guid? nextId = null;
-			if (nextMatch.Success && Guid.TryParse(nextMatch.Groups[2].Value, out var next))
-			{
-				nextId = next;
-				output = output.Substring(0, nextMatch.Groups[1].Index);
-			}
-
-			return (number1, number2, output, nextId);
-		}
-
-		private static RowKind GuessRowType(IReadOnlyList<(int? Number1, int? Number2, string Output, Guid? Next)> rowInfos)
-		{
-			if (rowInfos.All(x => x.Number1 is null && x.Number2 is null))
-				return RowKind.NoNumbers;
-
-			if (rowInfos.All(x => x.Number1 is not null && x.Number2 is null))
-			{
-				bool isSequential = true;
-				for (int i = 0; i < rowInfos.Count - 1; i++)
-				{
-					if (rowInfos[i].Number1 != i + 1)
+					if (kind == RandomSourceKind.CardSequence)
 					{
-						isSequential = false;
-						break;
+						var match = cardConfigRegex.Match(part);
+						if (!match.Success)
+							throw new FormatException($"Random does not have the expected format: {input}");
+						if (!Enum.TryParse<CardSourceConfiguration>(match.Groups[1].Value, out var config))
+							throw new FormatException($"Random does not have the expected format: {input}");
+						return EnumerableUtility.Enumerate((int) config);
 					}
+					else
+					{
+						var match = diceConfigRegex.Match(part);
+						if (!match.Success)
+							throw new FormatException($"Random does not have the expected format: {input}");
+						var countValue = match.Groups[1].Value;
+						var count = countValue.Length == 0 ? 1 : int.Parse(countValue);
+						var sides = int.Parse(match.Groups[2].Value);
+						return Enumerable.Repeat(sides, count);
+					}
+				})
+				.AsReadOnlyList();
+
+			return new RandomPlan(kind, configurations);
+		}
+
+		private static (RandomValueBase Value1, RandomValueBase? Value2, IReadOnlyList<int>? GuessedConfigs, string Output) CreateRowInfo(string line, RandomSourceBase? randomSource)
+		{
+			var regex = new Regex(@"^(?:((?:\d+,\s*)+\d+|\d+\s*-\s*\d+|\d+)(?:\s*\.\s+|\s*\:\s+|\s+))?(.*)$");
+			var match = regex.Match(line);
+			var value = match.Groups[1].Success ? match.Groups[1].Value : "";
+
+			if (!match.Groups[2].Success)
+				throw new FormatException($"Row output is missing: {line}");
+
+			var valueParts = value.Split('-', StringSplitOptions.TrimEntries | StringSplitOptions.RemoveEmptyEntries);
+			if (valueParts.Length > 2)
+				throw new FormatException($"Row value has more than one range: {line}");
+
+			RandomValueBase? value1 = null;
+			RandomValueBase? value2 = null;
+			IReadOnlyList<int>? configs = null;
+			if (randomSource is not null)
+			{
+				if (valueParts.Length > 0)
+					value1 = randomSource.TryConvertToValue(valueParts[0]);
+				if (valueParts.Length > 1)
+					value2 = randomSource.TryConvertToValue(valueParts[1]);
+			}
+			else
+			{
+				if (valueParts.Length > 0)
+				{
+					(value1, configs) = GuessValue(valueParts[0]);
+					if (value1 is null)
+						throw new FormatException($"First row value could not be parsed: {line}");
 				}
-				if (!isSequential)
-					return RowKind.Weighted;
+				if (valueParts.Length > 1)
+				{
+					(value2, var checkConfigs) = GuessValue(valueParts[1]);
+					if (value2 is null)
+						throw new FormatException($"Second row value could not be parsed: {line}");
+					if (value1!.Kind != value2.Kind)
+						throw new FormatException($"Row values must be the same kind: {line}");
+					if (value1!.Values.Count != value2.Values.Count)
+						throw new FormatException($"Row values must have the same number of values: {line}");
+					configs = configs!
+						.Zip(checkConfigs)
+						.Select(x =>
+						{
+							var (config1, config2) = x;
+							var mergedConfig = value1.Kind == RandomValueKind.Die ?
+								DieUtility.MergeConfigurations(config1, config2) :
+								(int?) CardUtility.MergeConfigurations((CardSourceConfiguration) config1, (CardSourceConfiguration) config2);
+							if (mergedConfig is null)
+								throw new FormatException($"Row values must have the same configuration: {line}");
+							return mergedConfig.Value;
+						})
+						.AsReadOnlyList();
+				}
+				if (valueParts.Length > 1 && (value1?.Values.Count > 1 || value2?.Values.Count > 1))
+					throw new FormatException($"Row value doesn't support a range and a list: {line}");
+			}
+			if (value1 is null)
+				value1 = new DieValue(1);
+
+			return (value1, value2, configs, match.Groups[2].Value);
+		}
+
+		private static (RandomValueBase Value, IReadOnlyList<int> Configs) GuessValue(string input)
+		{
+			var tokens = input.Split(new[] { ' ', ',' }, StringSplitOptions.RemoveEmptyEntries);
+
+			var values = new List<int>();
+			var configs = new List<int>();
+			RandomValueKind? kind = null;
+
+			for (int index = 0; index < tokens.Length; index++)
+			{
+				var token = tokens[index];
+				var (value, config) = RandomValueBase.TryParseSingleValue(token);
+				if (value is null)
+					throw new FormatException($"Value could not be parsed: {input}");
+				if (kind is null)
+					kind = value.Kind;
+				else if (kind != value.Kind)
+					throw new FormatException($"Value kind does not match previous values: {input}");
+
+				values.Add(value.Values[0]);
+				configs.Add(config!.Value);
 			}
 
-			if (rowInfos.All(x => x.Number1 is not null))
-				return RowKind.MixedRanges;
-
-			return RowKind.Unknown;
-		}
-
-		private static (RandomSourceData RandomSource, IReadOnlyList<RowData> Rows) CreateNoNumbersRowDatas(IReadOnlyList<(int? Number1, int? Number2, string Output, Guid? Next)> rowInfos)
-		{
-			var dice = GetBestDice(rowInfos.Count);
-			if (dice.Count != 1)
-				throw new NotImplementedException("Multiple dice not supported yet.");
-
-			var increment = dice.Count == 1 && dice[0] % rowInfos.Count == 0 ? dice[0] / rowInfos.Count : 1;
-			var randomSource = new RandomSourceData
+			return kind switch
 			{
-				Kind = RandomSourceKind.Die,
-				Dice = dice,
+				RandomValueKind.Die => (new DieValue(values), configs),
+				RandomValueKind.Card => (new CardValue(values), configs),
+				_ => throw new InvalidOperationException("Unknown random value kind"),
 			};
-
-			int startValue = 1;
-			var rows = rowInfos
-				.Select((info, index) =>
-				{
-					var data = new RowData
-					{
-						Min = startValue,
-						Max = startValue + increment - 1,
-						Output = info.Output,
-						Next = info.Next,
-					};
-					startValue += increment;
-					return data;
-				})
-				.AsReadOnlyList();
-
-			return (randomSource, rows);
 		}
 
-		private static (RandomSourceData RandomSource, IReadOnlyList<RowData> Rows) CreateMixedRangesRowDatas(IReadOnlyList<(int? Number1, int? Number2, string Output, Guid? Next)> rowInfos)
-		{
-			var randomSource = new RandomSourceData
-			{
-				Kind = RandomSourceKind.Die,
-				Dice = new[] { (rowInfos[^1].Number2 ?? rowInfos[^1].Number1)!.Value },
-			};
-
-			int lastMax = 0;
-			var rows = rowInfos
-				.Select((info, index) =>
-				{
-					if (info.Number1!.Value != lastMax + 1)
-						throw new InvalidOperationException("Dice values must be sequential.");
-					lastMax = (info.Number2 ?? info.Number1)!.Value;
-
-					var data = new RowData
-					{
-						Min = info.Number1!.Value,
-						Max = lastMax,
-						Output = info.Output,
-						Next = info.Next,
-					};
-					return data;
-				})
-				.AsReadOnlyList();
-
-			return (randomSource, rows);
-		}
-
-		private static (RandomSourceData RandomSource, IReadOnlyList<RowData> Rows) CreateWeightedRowDatas(IReadOnlyList<(int? Number1, int? Number2, string Output, Guid? Next)> rowInfos)
+		/*
+		private static (RandomSourceData RandomSource, IReadOnlyList<RowDataDto> Rows) CreateWeightedRowDatas(IReadOnlyList<(int? Number1, int? Number2, string Output, Guid? Next)> rowInfos)
 		{
 			var totalWeight = rowInfos.Sum(x => x.Number1!.Value);
 
@@ -285,7 +478,7 @@ namespace Oraculum.Data
 			var rows = rowInfos
 				.Select((info, index) =>
 				{
-					var data = new RowData
+					var data = new RowDataDto
 					{
 						Min = startValue,
 						Max = startValue + (info.Number1!.Value * increment) - 1,
@@ -299,6 +492,7 @@ namespace Oraculum.Data
 
 			return (randomSource, rows);
 		}
+		*/
 
 		private static IReadOnlyList<int> GetBestDice(int rowCount)
 		{
@@ -330,12 +524,6 @@ namespace Oraculum.Data
 			ReadingRows,
 		};
 
-		private enum RowKind
-		{
-			Unknown,
-			NoNumbers,
-			MixedRanges,
-			Weighted,
-		};
+		private static ILogSource Log { get; } = LogManager.CreateLogSource(nameof(DataImportUtility));
 	}
 }
