@@ -1,31 +1,57 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Windows.Threading;
 using Faithlife.Data;
 using Faithlife.Data.SqlFormatting;
 using GoldenAnvil.Utility;
 using GoldenAnvil.Utility.Logging;
+using GoldenAnvil.Utility.Windows.Async;
 using Microsoft.Data.Sqlite;
+using Microsoft.VisualStudio.Threading;
 using ProtoBuf;
 
 namespace Oraculum.Data
 {
-	public sealed class DataManager
+	public sealed partial class DataManager : IDisposable
 	{
-		public DataManager() { }
+		public DataManager()
+		{
+			m_dispatcher = Dispatcher.CurrentDispatcher;
+			m_taskGroup = new TaskGroup();
+			m_tableCache = new ConcurrentDictionary<Guid, TableMetadataImpl>();
+			m_tableReferencesCache = new ConcurrentDictionary<Guid, TableReference>();
+			m_workQueue = new BlockingCollection<Func<TaskStateController, Task>>();
+			m_workWatcher = TaskWatcher.Create(ProcessWorkQueueAsync, m_taskGroup);
+		}
 
-		public async Task InitializeAsync(CancellationToken cancellationToken)
+		public async Task InitializeAsync(TaskStateController state)
 		{
 			using var connector = CreateConnector();
 			var hasInfoTable = await connector.Command("select name from sqlite_master where type='table' AND name='Info'")
-				.QueryFirstOrDefaultAsync<string>(cancellationToken).ConfigureAwait(false) is not null;
+				.QueryFirstOrDefaultAsync<string>(state.CancellationToken).ConfigureAwait(false) is not null;
 			if (hasInfoTable)
-				await VerifyDbVersionAsync(connector, cancellationToken).ConfigureAwait(false);
+				await VerifyDbVersionAsync(connector, state.CancellationToken).ConfigureAwait(false);
 			else
-				await CreateDbAsync(connector, cancellationToken).ConfigureAwait(false);
+				await CreateDbAsync(connector, state.CancellationToken).ConfigureAwait(false);
+
+			var tableReferences = await connector.Command("select TableId, Title from TableMetadata")
+				.QueryAsync<(string id, string title)>(state.CancellationToken)
+				.ConfigureAwait(false);
+			foreach (var (idString, title) in tableReferences)
+			{
+				var id = Guid.Parse(idString);
+				m_tableReferencesCache.TryAdd(id, new TableReferenceImpl(id, title, this));
+			}
+
+			await state.ToSyncContext();
+
+			m_workWatcher.Start();
 		}
 
 		public async Task<IReadOnlyList<(string Key, string Value)>> GetAllSettingsAsync(CancellationToken cancellationToken)
@@ -37,29 +63,96 @@ namespace Oraculum.Data
 			return settings;
 		}
 
-		public async Task<string> GetSettingAsync(string key, CancellationToken cancellationToken)
+		public void SetSetting(string key, string value)
 		{
-			using var connector = CreateConnector();
-			return await connector.Command(Sql.Format($"select Value from Settings where Key = {key}"))
-				.QuerySingleAsync<string>(cancellationToken)
-				.ConfigureAwait(false);
+			m_workQueue.Add(async state =>
+			{
+				using var connector = CreateConnector();
+				await connector.Command(Sql.Format($"insert into Settings (Key, Value) values ({key}, {value}) on conflict do update set Value={value}"))
+					.ExecuteAsync(state.CancellationToken)
+					.ConfigureAwait(false);
+			});
 		}
 
-		public async Task SetSettingAsync(string key, string value, CancellationToken cancellationToken)
+		public void DeleteSetting(string key)
 		{
-			using var connector = CreateConnector();
-			await connector.Command(Sql.Format($"insert into Settings (Key, Value) values ({key}, {value}) on conflict do update set Value={value}"))
-				.ExecuteAsync(cancellationToken)
-				.ConfigureAwait(false);
+			m_workQueue.Add(async state =>
+			{
+				using var connector = CreateConnector();
+				await connector.Command(Sql.Format($"delete from Settings where Key = {key}"))
+					.ExecuteAsync(state.CancellationToken)
+					.ConfigureAwait(false);
+			});
 		}
 
-		public async Task DeleteSettingAsync(string key, CancellationToken cancellationToken)
+		public TableReference? GetTableReference(Guid id) =>
+			m_tableReferencesCache.TryGetValue(id, out var value) ? value : null;
+
+		public HashSet<string> GetAllTableTitles() =>
+			m_tableReferencesCache.Values.Select(x => x.Title).ToHashSet();
+
+		public async Task WaitForWriteAsync()
 		{
-			using var connector = CreateConnector();
-			await connector.Command(Sql.Format($"delete from Settings where Key = {key}"))
-				.ExecuteAsync(cancellationToken)
-				.ConfigureAwait(false);
+			m_workQueue.CompleteAdding();
+			await m_workWatcher.Task.ConfigureAwait(false);
 		}
+
+		public void Dispose()
+		{
+			DisposableUtility.Dispose(ref m_taskGroup);
+		}
+
+		private void UpdateTableTitle(Guid id, string title) =>
+			UpdateTableMetadata(id, Sql.Format($"Title={title}"));
+
+		private void UpdateTableSource(Guid id, string? source) =>
+			UpdateTableMetadata(id, Sql.Format($"Source={source}"));
+
+		private void UpdateTableAuthor(Guid id, string? author) =>
+			UpdateTableMetadata(id, Sql.Format($"Author={author}"));
+
+		private void UpdateTableRandomPlan(Guid id, RandomPlan randomPlan)
+		{
+			using var stream = new MemoryStream();
+			Serializer.Serialize(stream, randomPlan);
+			UpdateTableMetadata(id, Sql.Format($"RandomPlan={stream.ToArray()}"));
+		}
+
+		private void UpdateTableGroups(Guid id, IReadOnlyList<string> groups)
+		{
+			using var stream = new MemoryStream();
+			Serializer.Serialize(stream, groups);
+			UpdateTableMetadata(id, Sql.Format($"Groups={stream.ToArray()}"));
+		}
+
+		private void UpdateTableMetadata(Guid id, Sql updateSql)
+		{
+			m_workQueue.Add(async state =>
+			{
+				using var connector = CreateConnector();
+
+				var tableRowId = await connector
+					.Command(Sql.Format($@"update TableMetadata set {updateSql}, Modifed={GetCurrentDateOnlyString()}, Version=Version+1 where TableId={id}"))
+					.ExecuteAsync(state.CancellationToken)
+					.ConfigureAwait(false);
+			});
+		}
+
+		private static string GetCurrentDateOnlyString() =>
+			DateOnly.FromDateTime(DateTime.Now).ToString("yyyy-MM-dd", CultureInfo.InvariantCulture);
+
+		private async Task ProcessWorkQueueAsync(TaskStateController state)
+		{
+			await state.ToThreadPool();
+
+			while (!m_workQueue.IsCompleted)
+			{
+				if (m_workQueue.TryTake(out var work, Timeout.Infinite, state.CancellationToken))
+					await work(state).ConfigureAwait(false);
+			}
+		}
+
+		// Redo everything below
 
 		public async Task<IReadOnlyList<SetMetadata>> GetAllSetMetadataAsync(CancellationToken cancellationToken)
 		{
@@ -90,55 +183,76 @@ namespace Oraculum.Data
 				return await GetAllTableMetadatasAsync(cancellationToken).ConfigureAwait(false);
 
 			using var connector = CreateConnector();
-			var metadataStreams = await connector.Command(Sql.Format($@"
-				select tm.Metadata from TableMetadata tm
+			var tableIds = await connector.Command(Sql.Format($@"
+				select tm.TableId from TableMetadata tm
 				join SetTables st on st.TableRecordId = tm.RecordId
 				join SetMetadata sm on sm.RecordId = st.SetRecordId
 				where sm.SetId = {setId}
-			")).QueryAsync<byte[]>(cancellationToken).ConfigureAwait(false);
-			var metadatas = metadataStreams
-				.Select(x => Serializer.Deserialize<TableMetadata>(new ReadOnlySpan<byte>(x)))
-				.AsReadOnlyList();
+			")).QueryAsync<string>(cancellationToken).ConfigureAwait(false);
 
-			return metadatas;
-		}
-
-		public async Task<TableMetadata?> GetTableMetadataAsync(Guid tableId, CancellationToken cancellationToken)
-		{
-			using var connector = CreateConnector();
-
-			var sql = Sql.Format($"select Metadata from TableMetadata where TableId = {tableId}");
-			var metadata = await connector.Command(sql)
-				.QuerySingleOrDefaultAsync<byte[]?>(cancellationToken).ConfigureAwait(false);
-			return metadata is null ? null : Serializer.Deserialize<TableMetadata>(new ReadOnlySpan<byte>(metadata));
+			return await GetTableMetadatasAsync(tableIds.Select(x => Guid.Parse(x)), cancellationToken).ConfigureAwait(false);
 		}
 
 		public async Task<IReadOnlyList<TableMetadata>> GetTableMetadatasAsync(IEnumerable<Guid> tableIds, CancellationToken cancellationToken)
 		{
-			using var connector = CreateConnector();
+			List<TableMetadata> tables = new List<TableMetadata>();
+			List<Guid> uncachedTableIds = new List<Guid>();
 
-			var sql = Sql.Format($"select Metadata from TableMetadata where TableId in ({tableIds}...)");
-			var metadataStreams = await connector.Command(sql)
-				.QueryAsync<byte[]?>(cancellationToken).ConfigureAwait(false);
-			var metadatas = metadataStreams
-				.Select(x => Serializer.Deserialize<TableMetadata>(new ReadOnlySpan<byte>(x)))
-				.AsReadOnlyList();
-			return metadatas;
+			foreach (var tableId in tableIds)
+			{
+				if (m_tableCache.TryGetValue(tableId, out var table))
+					tables.Add(table);
+				else
+					uncachedTableIds.Add(tableId);
+			}
+
+			if (uncachedTableIds.Count != 0)
+			{
+				using var connector = CreateConnector();
+
+				var sql = Sql.Format($"select TableId, Title, Source, Author, Version, Created, Modified, Description, RandomPlan, Groups from TableMetadata where TableId in ({uncachedTableIds}...)");
+				var values = await connector.Command(sql)
+					.QueryAsync<(string TableId, string Title, string? Source, string? Author, long Version, string Created, string Modified, string? Description, byte[] RandomPlanBytes, byte[]? GroupsBytes)>(cancellationToken)
+					.ConfigureAwait(false);
+				var metadataDtos = values
+					.Select(x =>
+					{
+						var randomPlan = Serializer.Deserialize<RandomPlan>(new ReadOnlySpan<byte>(x.RandomPlanBytes));
+						var groups = Serializer.Deserialize<IReadOnlyList<string>>(new ReadOnlySpan<byte>(x.GroupsBytes));
+						return new TableMetadataDto(
+							tableId: Guid.Parse(x.TableId),
+							title: x.Title,
+							source: x.Source,
+							author: x.Author,
+							version: (int) x.Version,
+							created: DateOnly.ParseExact(x.Created, c_dateFormat, CultureInfo.InvariantCulture),
+							modified: DateOnly.ParseExact(x.Modified, c_dateFormat, CultureInfo.InvariantCulture),
+							description: x.Description,
+							randomPlan: randomPlan,
+							groups: groups
+							);
+					});
+
+				foreach (var metadataDto in metadataDtos)
+				{
+					var tableReference = (TableReferenceImpl) GetTableReference(metadataDto.TableId)!;
+					tables.Add(m_tableCache.GetOrAdd(tableReference!.Id, new TableMetadataImpl(tableReference, metadataDto, this)));
+				}
+			}
+
+			return tables;
 		}
 
-		public async Task<IReadOnlyList<TableMetadata>> GetAllTableMetadatasAsync(CancellationToken cancellationToken)
+		private async Task<IReadOnlyList<TableMetadata>> GetAllTableMetadatasAsync(CancellationToken cancellationToken)
 		{
 			using var connector = CreateConnector();
 
-			var metadataStreams = await connector.Command("select Metadata from TableMetadata").QueryAsync<byte[]>(cancellationToken).ConfigureAwait(false);
-			var metadatas = metadataStreams
-				.Select(x => Serializer.Deserialize<TableMetadata>(new ReadOnlySpan<byte>(x)))
-				.AsReadOnlyList();
+			var tableIds = await connector.Command("select TableId from TableMetadata").QueryAsync<string>(cancellationToken).ConfigureAwait(false);
 
-			return metadatas;
+			return await GetTableMetadatasAsync(tableIds.Select(x => Guid.Parse(x)), cancellationToken).ConfigureAwait(false);
 		}
 
-		public async Task<IReadOnlyList<RowData>> GetRowsAsync(Guid tableId, CancellationToken cancellationToken)
+		public async Task<IReadOnlyList<RowDataDto>> GetRowsAsync(Guid tableId, CancellationToken cancellationToken)
 		{
 			using var connector = CreateConnector();
 
@@ -148,38 +262,33 @@ namespace Oraculum.Data
 				where tm.TableId = {tableId}
 			")).QueryAsync<byte[]>(cancellationToken).ConfigureAwait(false);
 			var metadatas = metadataStreams
-				.Select(x => Serializer.Deserialize<RowData>(new ReadOnlySpan<byte>(x)))
+				.Select(x => Serializer.Deserialize<RowDataDto>(new ReadOnlySpan<byte>(x)))
 				.AsReadOnlyList();
 
 			return metadatas;
 		}
 
-		public async Task<HashSet<string>> GetAllTableTitlesAsync(CancellationToken cancellationToken)
-		{
-			using var connector = CreateConnector();
-
-			var metadataStreams = await connector.Command("select Metadata from TableMetadata").QueryAsync<byte[]>(cancellationToken).ConfigureAwait(false);
-			return metadataStreams
-				.Select(x => Serializer.Deserialize<TableMetadata>(new ReadOnlySpan<byte>(x)).Title)
-				.ToHashSet();
-		}
-
-		public async Task AddTableAsync(TableMetadata metadata, IEnumerable<RowData> rows, CancellationToken cancellationToken)
+		public async Task<TableReference> AddTableAsync(TableMetadataDto metadata, IEnumerable<RowDataDto> rows, CancellationToken cancellationToken)
 		{
 			using var connector = CreateConnector();
 
 			using var transaction = await connector.BeginTransactionAsync(cancellationToken).ConfigureAwait(false);
-			await AddTableImplAsync(connector, metadata, rows, cancellationToken).ConfigureAwait(false);
+			var tableReference = await AddTableImplAsync(connector, metadata, rows, cancellationToken).ConfigureAwait(false);
 			await connector.CommitTransactionAsync(cancellationToken).ConfigureAwait(false);
+
+			return tableReference;
 		}
 
-		private static async Task<long> AddTableImplAsync(DbConnector connector, TableMetadata metadata, IEnumerable<RowData> rows, CancellationToken cancellationToken)
+		private async Task<TableReference> AddTableImplAsync(DbConnector connector, TableMetadataDto metadata, IEnumerable<RowDataDto> rows, CancellationToken cancellationToken)
 		{
-			using var stream = new MemoryStream();
-			Serializer.Serialize(stream, metadata);
+			var tableReference = new TableReferenceImpl(metadata.TableId, metadata.Title, this);
+			m_tableReferencesCache.TryAdd(metadata.TableId, tableReference);
 
-			var dbId = await connector.Command(Sql.Format($@"
-					insert into TableMetadata (Metadata, TableId) values ({stream.ToArray()}, {metadata.Id}) returning RecordId
+
+			var randomPlan = CreateBytes(metadata.RandomPlan);
+			var groups = CreateBytes(metadata.Groups);
+			var tableRowId = await connector.Command(Sql.Format($@"
+					insert into TableMetadata (TableId, Title, Source, Author, Version, Created, Modified, Description, RandomPlan, Groups) values ({metadata.TableId}, {metadata.Title}, {metadata.Source}, {metadata.Author}, {metadata.Version}, {metadata.Created.ToString(c_dateFormat, CultureInfo.InvariantCulture)}, {metadata.Modified.ToString(c_dateFormat, CultureInfo.InvariantCulture)}, {metadata.Description}, {randomPlan}, {groups}) returning RecordId
 				")).QuerySingleAsync<long>(cancellationToken).ConfigureAwait(false);
 
 			foreach (var row in rows)
@@ -188,11 +297,18 @@ namespace Oraculum.Data
 				Serializer.Serialize(rowStream, row);
 
 				await connector.Command(Sql.Format($@"
-						insert into RowData (TableRecordId, Data) values ({dbId}, {rowStream.ToArray()})
+						insert into RowData (TableRecordId, Data) values ({tableRowId}, {rowStream.ToArray()})
 					")).ExecuteAsync(cancellationToken).ConfigureAwait(false);
 			}
 
-			return dbId;
+			return tableReference;
+
+			static byte[] CreateBytes<T>(T value)
+			{
+				using var stream = new MemoryStream();
+				Serializer.Serialize(stream, value);
+				return stream.ToArray();
+			}
 		}
 
 		private static DbConnector CreateConnector() =>
@@ -217,16 +333,11 @@ namespace Oraculum.Data
 			{
 				while (version < c_dbVersion)
 				{
+					/*
 					using (await connector.BeginTransactionAsync(cancellationToken).ConfigureAwait(false))
 					{
 						switch (version)
 						{
-						case 2:
-							await UpdateDatabaseFrom2To3Async(connector, cancellationToken).ConfigureAwait(false);
-							break;
-						case 1:
-							await UpdateDatabaseFrom1To2Async(connector, cancellationToken).ConfigureAwait(false);
-							break;
 						default:
 							throw new NotImplementedException($"Missing implementation of upgrade from version {version}");
 						};
@@ -239,47 +350,9 @@ namespace Oraculum.Data
 
 						await connector.CommitTransactionAsync(cancellationToken).ConfigureAwait(false);
 					}
+					*/
 				}
 			}
-		}
-
-		private static async Task UpdateDatabaseFrom2To3Async(DbConnector connector, CancellationToken cancellationToken)
-		{
-			await connector.Command(@"
-				create table Settings (
-					Key string not null primary key,
-					Value string not null
-				);
-			").ExecuteAsync(cancellationToken).ConfigureAwait(false);
-		}
-
-		private static async Task UpdateDatabaseFrom1To2Async(DbConnector connector, CancellationToken cancellationToken)
-		{
-			await connector.Command(@"
-				create table SetMetadata (
-					RecordId integer primary key,
-					SetId string not null,
-					Metadata blob not null
-				);
-
-				create table TableMetadata (
-					RecordId integer primary key,
-					TableId string not null,
-					Metadata blob not null
-				);
-
-				create table SetTables (
-					SetRecordId integer not null,
-					TableRecordId integer not null
-				);
-
-				create index SetTables_SetTable on SetTables (SetRecordId, TableRecordId);
-
-				create table RowData (
-					TableRecordId integer not null,
-					Data blob not null
-				);
-			").ExecuteAsync(cancellationToken).ConfigureAwait(false);
 		}
 
 		private static async Task CreateDbAsync(DbConnector connector, CancellationToken cancellationToken)
@@ -348,7 +421,7 @@ namespace Oraculum.Data
 			SqlSyntax = SqlSyntax.Sqlite,
 		};
 
-		private const int c_dbVersion = 3;
+		private const int c_dbVersion = 4;
 
 		private const string c_createSql = @"
 			create table Info (
@@ -358,20 +431,28 @@ namespace Oraculum.Data
 			);
 
 			create table Settings (
-				Key string not null primary key,
-				Value string not null
+				Key text not null primary key,
+				Value text not null
 			);
 
 			create table SetMetadata (
 				RecordId integer primary key,
-				SetId string not null,
+				SetId text not null,
 				Metadata blob not null
 			);
 
 			create table TableMetadata (
 				RecordId integer primary key,
-				TableId string not null,
-				Metadata blob not null
+				TableId text not null,
+				Title text not null,
+				Source text,
+				Author text,
+				Version integer not null,
+				Created text not null,
+				Modified text not null,
+				Description text,
+				RandomPlan blob not null,
+				Groups blob not null
 			);
 
 			create table SetTables (
@@ -389,5 +470,13 @@ namespace Oraculum.Data
 
 		private static readonly Guid s_ironswornSet = new Guid("599d53df-5076-4f1e-af03-0abe36991eba");
 		private static readonly Guid s_ironswornCustomSet = new Guid("04e1a881-9650-4cbb-8781-9f0b31391f83");
+		private const string c_dateFormat = "yyyy-MM-dd";
+
+		private readonly Dispatcher m_dispatcher;
+		private readonly ConcurrentDictionary<Guid, TableMetadataImpl> m_tableCache;
+		private readonly ConcurrentDictionary<Guid, TableReference> m_tableReferencesCache;
+		private readonly BlockingCollection<Func<TaskStateController, Task>> m_workQueue;
+		private readonly TaskWatcher m_workWatcher;
+		private TaskGroup m_taskGroup;
 	}
 }
