@@ -1,23 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Windows.Threading;
 using GoldenAnvil.Utility;
-using GoldenAnvil.Utility.Windows;
-using GoldenAnvil.Utility.Windows.Async;
 
 namespace Oraculum.Data
 {
-	public sealed class SettingsManager : NotifyPropertyChangedDispatcherBase
+	
+	public sealed class SettingsManager
 	{
 		public SettingsManager()
 		{
 			m_preferences = new Dictionary<string, string>();
-			m_taskGroup = new TaskGroup();
-			m_writeTasks = new List<TaskWatcher>();
+			m_dispatcher = Dispatcher.CurrentDispatcher;
 		}
+
+		public event EventHandler<GenericEventArgs<string>>? SettingChanged;
 
 		public async Task InitializeAsync(CancellationToken cancellationToken)
 		{
@@ -41,59 +40,29 @@ namespace Oraculum.Data
 				Clear(key);
 
 			VerifyAccess();
-			var localKey = key;
 			var newValue = ConversionUtility.Convert<string>(value);
-			if (!m_preferences.TryGetValue(localKey, out var oldValue) || oldValue != newValue)
+			if (!m_preferences.TryGetValue(key, out var oldValue) || oldValue != newValue)
 			{
-				m_preferences[localKey] = newValue;
-				DoWork(async state =>
-				{
-					var data = AppModel.Instance.Data;
-					await data.SetSettingAsync(localKey, newValue, state.CancellationToken).ConfigureAwait(false);
-				});
+				m_preferences[key] = newValue;
+				AppModel.Instance.Data.SetSetting(key, newValue);
+				SettingChanged.Raise(this, new GenericEventArgs<string>(key));
 			}
 		}
 
 		public void Clear(string key)
 		{
 			VerifyAccess();
-			var localKey = key;
-			if (m_preferences.Remove(localKey))
-			{
-				DoWork(async state =>
-				{
-					var data = AppModel.Instance.Data;
-					await data.DeleteSettingAsync(localKey, state.CancellationToken).ConfigureAwait(false);
-				});
-			}
+			if (m_preferences.Remove(key))
+				AppModel.Instance.Data.DeleteSetting(key);
 		}
 
-		private void DoWork(Func<TaskStateController, Task> work)
+		private void VerifyAccess()
 		{
-			var watcher = TaskWatcher.Create(work, m_taskGroup);
-			m_writeTasks.Add(watcher);
-			watcher.PropertyChanged += OnWorkCompleted;
-			watcher.Start();
-
-			void OnWorkCompleted(object? sender, PropertyChangedEventArgs e)
-			{
-				if (e.PropertyName == nameof(TaskWatcher.IsCompleted))
-				{
-					m_writeTasks.Remove(watcher);
-					watcher.PropertyChanged -= OnWorkCompleted;
-				}
-			}
-		}
-
-		public async Task WaitForWriteAsync(CancellationToken cancellationToken)
-		{
-			var tasks = m_writeTasks.Select(x => x.Task).ToArray();
-			await Task.WhenAll(tasks).ConfigureAwait(false);
+			if (!m_dispatcher.CheckAccess())
+				throw new InvalidOperationException("This code must be called on the same dispatcher as the object was created.");
 		}
 
 		private readonly Dictionary<string, string> m_preferences;
-		private readonly List<TaskWatcher> m_writeTasks;
-
-		private TaskGroup m_taskGroup;
+		private readonly Dispatcher m_dispatcher;
 	}
 }
