@@ -1,13 +1,18 @@
-﻿using System;
+﻿using Oraculum.Data;
 
 namespace Oraculum.Engine
 {
 	public sealed class RollResult
   {
-    public Guid TableId { get; init; }
-    public string? TableTitle { get; init; }
-    public string? Key { get; init; }
-    public string? Output { get; init; }
-    public Guid? Next { get; init; }
+    public RollResult(TableReference table, string value, string output)
+    {
+			Table = table;
+			Key = value;
+			Output = output;
+		}
+
+    public TableReference Table { get; init; }
+    public string Key { get; init; }
+    public string Output { get; init; }
   }
 }
