@@ -1,11 +1,11 @@
 ﻿using System;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Documents;
 using GoldenAnvil.Utility;
 using GoldenAnvil.Utility.Windows;
-using GoldenAnvil.Utility.Windows.Async;
+using Oraculum.Data;
 using Oraculum.Engine;
+using Oraculum.UI;
 
 namespace Oraculum.ViewModels
 {
@@ -18,24 +18,24 @@ namespace Oraculum.ViewModels
 
 		public FlowDocument Document { get; }
 
-		public void RollStarted(Guid tableId, string tableTitle)
+		public void RollStarted(TableReference table)
 		{
-			if (tableId != m_lastRollResultTableId)
+			if (table.Id != m_lastRollResultTableId)
 			{
 				var tableParagraph = new Paragraph().WithStyle("RollResultTableParagraphStyle");
 				tableParagraph.Inlines.AddRange(TextElementUtility.FormatInlineString(
 					OurResources.RollResultTable,
 					null,
-					new Hyperlink(new Run(tableTitle).WithStyle("RollResultTableTitleRunStyle"))
+					new Hyperlink(new Run(table.Title).WithStyle("RollResultTableTitleRunStyle"))
 					{
-						Command = new DelegateCommand(() => AppModel.Instance.OpenTable(tableId))
+						Command = new DelegateCommand(() => AppModel.Instance.OpenTable(table))
 					}
 					));
 				AddParagraph(tableParagraph, true);
 			}
 		}
 
-		public async Task AddAsync(TaskStateController state, RollResult result)
+		public void Add(RollResult result)
 		{
 			var paragraph = new Paragraph().WithStyle("RollResultParagraphStyle");
 			paragraph.Inlines.AddRange(TextElementUtility.FormatInlineString(
@@ -44,25 +44,11 @@ namespace Oraculum.ViewModels
 				new Run(result.Key).WithStyle("RollResultKeyRunStyle")
 				));
 			paragraph.Inlines.Add(new LineBreak());
-			paragraph.Inlines.Add(new Run(result.Output).WithStyle("RollResultOutputRunStyle"));
-
-			if (result.Next is not null)
-			{
-				var nextTableMetadata = await AppModel.Instance.Data.GetTableMetadataAsync(result.Next.Value, state.CancellationToken).ConfigureAwait(false);
-				paragraph.Inlines.Add(new LineBreak());
-				paragraph.Inlines.AddRange(TextElementUtility.FormatInlineString(
-					"also roll on {0}",
-					null,
-					new Hyperlink(new Run(nextTableMetadata?.Title ?? "<unknown>").WithStyle("RollResultOutputRunStyle"))
-					{
-						Command = new DelegateCommand(() => AppModel.Instance.OpenTable(result.Next.Value))
-					}
-					));
-			}
+			paragraph.Inlines.AddRange(TokenStringUtility.TokenStringToInlines(result.Output, "RollResultOutputRunStyle"));
 
 			AddParagraph(paragraph, true);
 
-			m_lastRollResultTableId = result.TableId;
+			m_lastRollResultTableId = result.Table.Id;
 		}
 
 		private void AddParagraph(Paragraph paragraph, bool shouldBringIntoView)
