@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.ComponentModel;
 using System.Threading.Tasks;
 using GoldenAnvil.Utility;
 using GoldenAnvil.Utility.Logging;
@@ -11,31 +11,31 @@ using Oraculum.Engine;
 
 namespace Oraculum.ViewModels
 {
-	public sealed class TableViewModel : TreeLeafBase
+	public sealed class TableViewModel : TreeLeafBase, IDisposable
 	{
 		public TableViewModel(TableMetadata metadata)
 		{
-			m_id = metadata.Id;
+			m_metadata = metadata;
+			m_metadata.PropertyChanged += OnMetadataPropertyChanged;
+
+			Title = metadata.Title;
 			m_author = metadata.Author ?? "";
 			m_version = metadata.Version;
 			m_created = metadata.Created;
 			m_modified = metadata.Modified;
 			m_groups = metadata.Groups;
 			m_randomPlan = metadata.RandomPlan;
+			m_randomSource = RandomSourceBase.Create(m_randomPlan);
+			m_valueGenerator = new ValueGenerator(metadata.TableReference, m_randomSource, OnValueGenerated);
 			m_useManualRoll = AppModel.Instance.Settings.Get<bool>(SettingsKeys.RollValueManually);
 			m_extraTables = new Dictionary<Guid, (TableMetadata Metadata, ValueToResultMapper ResultMapper)>();
 
-			Title = metadata.Title ?? "";
-
-			m_valueGenerators = new List<ValueGeneratorViewModelBase>();
-			ValueGenerators = CreateValueGenerators(m_randomPlan, m_useManualRoll);
+			AppModel.Instance.Settings.SettingChanged += OnSettingChanged;
 		}
 
-		public Guid Id
-		{
-			get => VerifyAccess(m_id);
-			set => SetPropertyField(value, ref m_id);
-		}
+		public ValueGenerator ValueGenerator => m_valueGenerator;
+
+		public TableReference TableReference => m_metadata.TableReference;
 
 		public string Author
 		{
@@ -46,31 +46,22 @@ namespace Oraculum.ViewModels
 		public int Version
 		{
 			get => VerifyAccess(m_version);
-			set => SetPropertyField(value, ref m_version);
 		}
 
-		public DateTime Created
+		public DateOnly Created
 		{
 			get => VerifyAccess(m_created);
-			set => SetPropertyField(value, ref m_created);
 		}
 
-		public DateTime Modified
+		public DateOnly Modified
 		{
 			get => VerifyAccess(m_modified);
-			set => SetPropertyField(value, ref m_modified);
 		}
 
 		public IReadOnlyList<string> Groups
 		{
 			get => VerifyAccess(m_groups);
 			set => SetPropertyField(value, ref m_groups);
-		}
-
-		public RollLogViewModel? RollLog
-		{
-			get => VerifyAccess(m_rollLog);
-			set => SetPropertyField(value, ref m_rollLog);
 		}
 
 		public bool IsWorking
@@ -85,33 +76,7 @@ namespace Oraculum.ViewModels
 			set
 			{
 				if (SetPropertyField(value, ref m_useManualRoll))
-				{
 					AppModel.Instance.Settings.Set(SettingsKeys.RollValueManually, value);
-					var generators = CreateValueGenerators(m_randomPlan, m_useManualRoll);
-					ValueGenerators = generators;
-				}
-			}
-		}
-
-		public IReadOnlyList<ValueGeneratorViewModelBase> ValueGenerators
-		{
-			get => VerifyAccess(m_valueGenerators);
-			private set
-			{
-				var oldGenerators = m_valueGenerators;
-				if (SetPropertyField(value, ref m_valueGenerators))
-				{
-					foreach (var generator in oldGenerators)
-					{
-						generator.RollStarted -= OnRollStarted;
-						generator.ValueGenerated -= OnValueGenerated;
-					}
-					foreach (var generator in m_valueGenerators)
-					{
-						generator.RollStarted += OnRollStarted;
-						generator.ValueGenerated += OnValueGenerated;
-					}
-				}
 			}
 		}
 
@@ -125,7 +90,7 @@ namespace Oraculum.ViewModels
 
 			await state.ToSyncContext();
 			IsWorking = true;
-			var tableId = Id;
+			var tableId = m_metadata.Id;
 			var data = AppModel.Instance.Data;
 
 			try
@@ -133,13 +98,13 @@ namespace Oraculum.ViewModels
 				await state.ToThreadPool();
 
 				var rows = await data.GetRowsAsync(tableId, state.CancellationToken).ConfigureAwait(false);
+
+				/*
 				var extraTableIds = rows
-					.Select(x => x.Next)
-					.WhereNotNull()
-					.Select(x => x!.Value)
+					.SelectMany(x => TokenStringUtility.GetTableReferences(x.Output).Select(x => x.Id))
 					.Distinct()
 					.AsReadOnlyList();
-				var extraTables = new List<(TableMetadata Metadata, IReadOnlyList<RowData> Rows)>();
+				var extraTables = new List<(TableMetadata Metadata, IReadOnlyList<RowDataDto> Rows)>();
 				if (extraTableIds.Count != 0)
 				{
 					var extraTableMetadatas = await data.GetTableMetadatasAsync(extraTableIds, state.CancellationToken).ConfigureAwait(false);
@@ -150,14 +115,19 @@ namespace Oraculum.ViewModels
 							extraTables.Add((extraTableMetadata, extraRows));
 					}
 				}
+				*/
+
 				await state.ToSyncContext();
 
-				m_resultMappers = m_valueGenerators.Select(x => new ValueToResultMapper(Id, Title, rows)).ToList();
+				m_resultMapper = new ValueToResultMapper(TableReference, m_randomSource, rows);
+
+				/*
 				foreach (var (extraMetadata, extraRows) in extraTables)
 				{
 					var mapper = new ValueToResultMapper(extraMetadata.Id, extraMetadata.Title, extraRows);
 					m_extraTables.Add(extraMetadata.Id, (extraMetadata, mapper));
 				}
+				*/
 
 				m_isLoaded = true;
 			}
@@ -168,70 +138,80 @@ namespace Oraculum.ViewModels
 			}
 		}
 
-		public void Roll()
+		public void Roll() => m_valueGenerator.Roll();
+
+		public void CommitChanges()
 		{
-			foreach (var generator in m_valueGenerators)
-				generator.Roll();
+			m_isUpdatingMetadata++;
+
+			m_metadata.Author = Author;
+			m_metadata.Groups = Groups;
+
+			m_isUpdatingMetadata--;
 		}
 
-		private void OnRollStarted(object? sender, EventArgs e)
+		public void Dispose()
 		{
-			if (m_valueGenerators.Where(x => x.IsRollStarted).Count() == 1)
-				m_rollLog?.RollStarted(m_id, Title);
+			m_metadata.PropertyChanged -= OnMetadataPropertyChanged;
+			AppModel.Instance.Settings.SettingChanged += OnSettingChanged;
+			DisposableUtility.Dispose(ref m_valueGenerator);
 		}
 
-		private void OnValueGenerated(object? sender, GenericEventArgs<RandomValueBase> e)
+		private void OnValueGenerated(RandomValueBase value)
 		{
-			if (m_valueGenerators.All(x => x.GeneratedValue is not null))
+			var result = m_resultMapper!.GetResult(value);
+			AppModel.Instance.RollLog!.Add(result);
+		}
+
+		private void OnMetadataPropertyChanged(object? sender, PropertyChangedEventArgs e)
+		{
+			if (m_isUpdatingMetadata > 0)
+				return;
+
+			switch (e.PropertyName)
 			{
-				var values = m_valueGenerators.Select(x => x.GeneratedValue!).AsReadOnlyList();
-				var results = new List<RollResult>(m_resultMappers!.Count);
-				for (int i = 0; i < m_resultMappers!.Count; i++)
-				{
-					var result = m_resultMappers[i].GetResult(values[i]);
-					results.Add(result);
-				}
-				TaskWatcher.Execute(async state =>
-				{
-					foreach (var result in results)
-						await m_rollLog!.AddAsync(state, result).ConfigureAwait(false);
-					await state.ToSyncContext();
-					foreach (var generator in m_valueGenerators)
-						generator.OnReportingFinished();
-				}, AppModel.Instance.TaskGroup);
+				case nameof(TableMetadata.Author):
+					Author = m_metadata.Author ?? "";
+					break;
+				case nameof(TableMetadata.Version):
+					SetPropertyField(nameof(Version), m_metadata.Version, ref m_version);
+					break;
+				case nameof(TableMetadata.Modified):
+				SetPropertyField(nameof(Modified), m_metadata.Modified, ref m_modified);
+					break;
+				case nameof(TableMetadata.Groups):
+					Groups = m_metadata.Groups;
+					break;
+				case nameof(TableMetadata.RandomPlan):
+					throw new NotImplementedException("Updating of RandomPlan is not yet implemented.");
+				default:
+					throw new InvalidOperationException($"Unexpected property name: {e.PropertyName}");
 			}
 		}
 
-		private static IReadOnlyList<ValueGeneratorViewModelBase> CreateValueGenerators(IReadOnlyList<RandomSourceData> randomPlan, bool rollManually)
+		private void OnSettingChanged(object? sender, GenericEventArgs<string> e)
 		{
-			List<ValueGeneratorViewModelBase> generators = new();
-
-			foreach (var sourceInfo in randomPlan)
-			{
-				var source = RandomSourceBase.Create(sourceInfo);
-				var generator = ValueGeneratorViewModelBase.Create(source, rollManually);
-				generators.Add(generator);
-			}
-
-			return generators;
+			if (e.Value == SettingsKeys.RollValueManually)
+				UseManualRoll = AppModel.Instance.Settings.Get<bool>(SettingsKeys.RollValueManually);
 		}
 
 		private static ILogSource Log { get; } = LogManager.CreateLogSource(nameof(TableViewModel));
 
+		private readonly TableMetadata m_metadata;
 		private readonly Dictionary<Guid, (TableMetadata Metadata, ValueToResultMapper ResultMapper)> m_extraTables;
 
-		private Guid m_id;
+		private int m_isUpdatingMetadata;
 		private string m_author;
 		private int m_version;
-		private DateTime m_created;
-		private DateTime m_modified;
+		private DateOnly m_created;
+		private DateOnly m_modified;
 		private IReadOnlyList<string> m_groups;
-		private IReadOnlyList<RandomSourceData> m_randomPlan;
+		private RandomPlan m_randomPlan;
+		private RandomSourceBase m_randomSource;
 		private bool m_isLoaded;
 		private bool m_isWorking;
-		private RollLogViewModel? m_rollLog;
-		private IReadOnlyList<ValueGeneratorViewModelBase> m_valueGenerators;
-		private IReadOnlyList<ValueToResultMapper>? m_resultMappers;
+		private ValueGenerator m_valueGenerator;
+		private ValueToResultMapper? m_resultMapper;
 		private bool m_useManualRoll;
 	}
 }
